astro/requirements.txt
dbt-postgres
apache-airflow-providers-postgres==5.10.0
psycopg2-binary==2.9.9